Added Unit Tests
These unit tests are really just place holders for now, a lot more will need to be added, and a lot more edge cases. Also another thing of note, the GL_Renderer will defenitely have to be split up in different classes somehow. Also the way captions are currently rendered is not expandable at all, which will have to be fixed. The box, polygon, and arrow buttons have been disabled since it'll be a while until they'll be implemented